[1/2] Rename Repository to IndexQueryService (#2277)
Rename module
